RFC 4193: don't lookup reverse lookups for  ULA addresses
RFC 4193 says reverse queries for _locally assigned_ IPv6 Local addresses
(FD00::/8) should not be forwarded to the root nameservers.  If dnscache finds
an appropriate nameserver under $ROOT/servers for such a query, it will ask it
and not answer with NXDOMAIN directly.

Similarly, in FORWARDONLY mode, those caches in '@' might know appropriate
nameservers for domains under d.f.ip6.arpa, so dnscache will forward the query
and not answer with NXDOMAIN directly.